+++ lib/flutter/homeassistant_ws_flutter.dart
@@ -2,3 +2,4 @@ library;
 
 export 'src/homeassistant_ws_builder.dart';
 export 'src/homeassistant_provider.dart';
+export '../homeassistant_ws.dart';