+++ CHANGELOG.md
@@ -1,3 +1,6 @@
 ## 1.0.0
 
-- Initial version.
+- Initial version of homeassistant_ws.
+- Features:
+    - Listen for state changes on different entities.
+    - Template widget for integration flutter using ChangeNotifier & Provider.